# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "DataQuality": {
        "additionalProperties": false,
        "properties": {
          "data_label": {
            "title": "Data Label",
            "type": "string"
          },
          "database": {
            "$ref": "#/components/schemas/DatabaseInfo"
          },
          "health": {
            "items": {
              "$ref": "#/components/schemas/SourceHealth"
            },
            "title": "Health",
            "type": "array"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/IngestionRun"
            },
            "title": "Runs",
            "type": "array"
          },
          "server_time": {
            "format": "date-time",
            "title": "Server Time",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceState"
            },
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "server_time",
          "data_label",
          "database",
          "sources",
          "health",
          "runs",
          "notes"
        ],
        "title": "DataQuality",
        "type": "object"
      },
...
      "EventItem": {
        "additionalProperties": false,
        "properties": {
          "actual": {
            "title": "Actual",
            "type": "number"
          },
          "data_class": {
            "default": "SIMULATED",
            "enum": [
              "LIVE",
              "NEAR-REAL-TIME",
              "RECENT",
              "HISTORICAL",
              "PREDICTED",
              "SIMULATED",
              "STATIC"
            ],
            "title": "Data Class",
            "type": "string"
          },
          "detected_at": {
            "format": "date-time",
            "title": "Detected At",
            "type": "string"
          },
          "end": {
            "format": "date-time",
            "title": "End",
            "type": "string"
          },
          "expected": {
            "title": "Expected",
            "type": "number"
          },
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "surge",
              "drop"
            ],
            "title": "Kind",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "severity": {
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Severity",
            "type": "string"
          },
          "start": {
            "format": "date-time",
            "title": "Start",
            "type": "string"
          },
          "zone": {
            "title": "Zone",
            "type": "string"
          },
          "zone_id": {
            "title": "Zone Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "zone_id",
          "zone",
          "kind",
          "severity",
          "start",
          "end",
          "actual",
          "expected",
          "score",
          "detected_at",
          "explanation"
        ],
        "title": "EventItem",
        "type": "object"
      },
...
      "Geometry": {
        "additionalProperties": false,
        "properties": {
          "attribution": {
            "title": "Attribution",
            "type": "string"
          },
          "bbox": {
            "items": {
              "type": "number"
            },
            "title": "Bbox",
            "type": "array"
          },
          "city": {
            "title": "City",
            "type": "string"
          },
          "licence": {
            "title": "Licence",
            "type": "string"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "zones": {
            "items": {
              "$ref": "#/components/schemas/ZoneGeometry"
            },
            "title": "Zones",
            "type": "array"
          }
        },
        "required": [
          "city",
          "bbox",
          "attribution",
          "licence",
          "method",
          "zones"
        ],
        "title": "Geometry",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngestionRun": {
        "additionalProperties": false,
        "properties": {
          "duration_ms": {
            "title": "Duration Ms",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "format": "date-time",
            "title": "Finished At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "records_in": {
            "title": "Records In",
            "type": "integer"
          },
          "records_ok": {
            "title": "Records Ok",
            "type": "integer"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source",
          "started_at",
          "finished_at",
          "ok",
          "records_in",
          "records_ok",
          "duration_ms"
        ],
        "title": "IngestionRun",
        "type": "object"
      },
...
      "SourceState": {
        "additionalProperties": false,
        "properties": {
          "age_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Since the source's own timestamp.",
            "title": "Age S"
          },
          "consecutive_failures": {
            "default": 0,
            "title": "Consecutive Failures",
            "type": "integer"
          },
          "data_class": {
            "enum": [
              "LIVE",
              "NEAR-REAL-TIME",
              "RECENT",
              "HISTORICAL",
              "PREDICTED",
              "SIMULATED",
              "STATIC"
            ],
            "title": "Data Class",
            "type": "string"
          },
          "disabled_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disabled Reason"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "freshness": {
            "enum": [
              "LIVE",
              "DELAYED",
              "STALE",
              "OFFLINE",
              "DISABLED",
              "NOT_PERIODIC"
            ],
            "title": "Freshness",
            "type": "string"
          },
          "interval_s": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interval S"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "last_observed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Observed At"
          },
          "last_success_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Success At"
          },
          "licence": {
            "title": "Licence",
            "type": "string"
          },
          "modelled": {
            "title": "Modelled",
            "type": "boolean"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "records_total": {
            "default": 0,
            "title": "Records Total",
            "type": "integer"
          },
          "runs": {
            "default": 0,
            "title": "Runs",
            "type": "integer"
          },
          "since_poll_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Since our last good poll.",
            "title": "Since Poll S"
          },
          "successes": {
            "default": 0,
            "title": "Successes",
            "type": "integer"
          }
        },
        "required": [
          "key",
          "label",
          "provider",
          "data_class",
          "modelled",
          "licence",
          "note",
          "enabled",
          "freshness"
        ],
        "title": "SourceState",
        "type": "object"
      },
      "StateSnapshot": {
        "additionalProperties": false,
        "properties": {
          "city": {
            "title": "City",
            "type": "string"
          },
          "data_label": {
            "title": "Data Label",
            "type": "string"
          },
          "demand_class": {
            "default": "SIMULATED",
            "enum": [
              "LIVE",
              "NEAR-REAL-TIME",
              "RECENT",
              "HISTORICAL",
              "PREDICTED",
              "SIMULATED",
              "STATIC"
            ],
            "title": "Demand Class",
            "type": "string"
          },
          "environment": {
            "$ref": "#/components/schemas/EnvironmentBlock"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventItem"
            },
            "title": "Events",
            "type": "array"
          },
          "forecast_made_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Made At"
          },
          "forecast_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Model"
          },
          "freshness": {
            "description": "The least healthy source that applies.",
            "enum": [
              "LIVE",
              "DELAYED",
              "STALE",
              "OFFLINE",
              "DISABLED",
              "NOT_PERIODIC"
            ],
            "title": "Freshness",
            "type": "string"
          },
          "selector": {
            "enum": [
              "now",
              "-15m",
              "-1h",
              "-6h",
              "today",
              "forecast"
            ],
            "title": "Selector",
            "type": "string"
          },
          "seq": {
            "description": "Increases whenever the operational store changes.",
            "title": "Seq",
            "type": "integer"
          },
          "server_time": {
            "format": "date-time",
            "title": "Server Time",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceState"
            },
            "title": "Sources",
            "type": "array"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "totals": {
            "$ref": "#/components/schemas/Totals"
          },
          "window_end": {
            "format": "date-time",
            "title": "Window End",
            "type": "string"
          },
          "window_note": {
            "title": "Window Note",
            "type": "string"
          },
          "window_start": {
            "format": "date-time",
            "title": "Window Start",
            "type": "string"
          },
          "worker": {
            "$ref": "#/components/schemas/WorkerState"
          },
          "zones": {
            "items": {
              "$ref": "#/components/schemas/ZoneValue"
            },
            "title": "Zones",
            "type": "array"
          }
        },
        "required": [
          "server_time",
          "selector",
          "window_start",
          "window_end",
          "window_note",
          "data_label",
          "city",
          "timezone",
          "seq",
          "freshness",
          "worker",
          "zones",
          "totals",
          "events",
          "environment",
          "sources"
        ],
        "title": "StateSnapshot",
        "type": "object"
      },
      "StreamStatus": {
        "additionalProperties": false,
        "properties": {
          "dropped_events": {
            "title": "Dropped Events",
            "type": "integer"
          },
          "max_streams": {
            "title": "Max Streams",
            "type": "integer"
          },
          "published": {
            "title": "Published",
            "type": "integer"
          },
          "streams": {
            "title": "Streams",
            "type": "integer"
          }
        },
        "required": [
          "streams",
          "max_streams",
          "published",
          "dropped_events"
        ],
        "title": "StreamStatus",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Read-only analytics, forecasts, anomaly events and simulated repositioning scenarios over NYC yellow-taxi demand. Sample mode serves TEST / SYNTHETIC DATA; the `X-Data-Label` response header always says which is being served. Optimization outputs are SIMULATED SCENARIOS under explicit assumptions.",
    "title": "MobilityOps API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/state/events": {
      "get": {
        "description": "Events the live rule found among today's completed hours (SIMULATED demand).",
        "operationId": "state_events_api_v1_state_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EventItem"
                  },
                  "title": "Response State Events Api V1 State Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "State Events",
        "tags": [
          "state"
        ]
      }
    },
    "/api/v1/state/geometry": {
      "get": {
        "description": "Zone polygons and attribution (OpenStreetMap, ODbL). Static: cache freely.",
        "operationId": "state_geometry_api_v1_state_geometry_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Geometry"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "State Geometry",
        "tags": [
          "state"
        ]
      }
    },
    "/api/v1/state/quality": {
      "get": {
        "description": "The data-quality centre: database checks, source health and recent ingestion runs.",
        "operationId": "state_quality_api_v1_state_quality_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataQuality"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "State Quality",
        "tags": [
          "state"
        ]
      }
    },
    "/api/v1/state/runs": {
      "get": {
        "operationId": "state_runs_api_v1_state_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IngestionRun"
                  },
                  "title": "Response State Runs Api V1 State Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "State Runs",
        "tags": [
          "state"
        ]
      }
    },
    "/api/v1/state/snapshot": {
      "get": {
        "description": "Every zone's simulated demand against its forecast at ``at`` (NOW, -15m, -1h, -6h,\ntoday, or forecast), plus events, weather, and the freshness of every source.",
        "operationId": "state_snapshot_api_v1_state_snapshot_get",
        "parameters": [
          {
            "in": "query",
            "name": "at",
            "required": false,
            "schema": {
              "default": "now",
              "enum": [
                "now",
                "-15m",
                "-1h",
                "-6h",
                "today",
                "forecast"
              ],
              "title": "At",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "State Snapshot",
        "tags": [
          "state"
        ]
      }
    },
    "/api/v1/state/sources": {
      "get": {
        "description": "Every data source: what it is, its licence, and its freshness right now.",
        "operationId": "state_sources_api_v1_state_sources_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SourceState"
                  },
                  "title": "Response State Sources Api V1 State Sources Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "State Sources",
        "tags": [
          "state"
        ]
      }
    },
    "/api/v1/state/stream": {
      "get": {
        "description": "Server-sent events: ``hello`` (a full snapshot), ``snapshot`` when the store changes\n(and at least every 30 s), and ``heartbeat`` every 10 s. ``limit`` ends the stream after\nthat many events (for tests and ``curl``).",
        "operationId": "state_stream_api_v1_state_stream_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 10000,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {}
            },
            "description": "Server-sent events"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "State Stream",
        "tags": [
          "state"
        ]
      }
    },
    "/api/v1/state/stream/status": {
      "get": {
        "operationId": "state_stream_status_api_v1_state_stream_status_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "State Stream Status",
        "tags": [
          "state"
        ]
      }
    },
    "/api/v1/state/zones/{zone_id}": {
      "get": {
        "operationId": "state_zone_api_v1_state_zones__zone_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "zone_id",
            "required": true,
            "schema": {
              "title": "Zone Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZoneDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "State Zone",
        "tags": [
          "state"
        ]
      }
    },
...
  }
}